alive-api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
email-validator
greenlet
